Relocates custom logic into 'deploymnets' package; Removes stale scripts that use legacy deployment methods
