+++ deployment/artifacts/lynx/lynx-alpha-13-child-registry.json
