+++ deployment/registry.py
@@ -6,7 +6,9 @@ from typing import Dict, List, NamedTuple, Optional
 from ape.contracts import ContractInstance
 from eth_typing import ChecksumAddress
 from eth_utils import to_checksum_address
-from scripts.utils import check_registry_filepath
+
+from deployment.params import get_contract_container
+from deployment.utils import check_registry_filepath
 
 
 class RegistryEntry(NamedTuple):
@@ -187,3 +189,14 @@ def merge_registries(
     print(f"Merged registry output to {output_filepath}")
 
     return output_filepath
+
+
+def contracts_from_registry(registry_filepath: Path):
+    registry_entries = read_registry(filepath=registry_filepath)
+    deployments = list()
+    for registry_entry in registry_entries:
+        contract_type = registry_entry.contract_name
+        contract_container = get_contract_container(contract_type)
+        contract_instance = contract_container.at(registry_entry.contract_address)
+        deployments.append(contract_instance)
+    return deployments
