+++ setup.cfg
@@ -8,3 +8,6 @@ include_trailing_comma = True
 line_length = 100
 multi_line_output = 3
 use_parentheses = True
+
+[options]
+packages = find:
